added javadocs and moved type declaration to corrct file
