+++ src/renderer/components/configure-recording-set-page/configure-recording-set.tsx
@@ -27,10 +27,9 @@ import AddRecordingSetButton from './add-recording-set-button';
 import CreatedRecordingSetList from './created-recording-set-list';
 import SetMetaConfiguration from './set-meta-configuration';
 import SetRecordingListConfiguration from './set-recording-list-configuration';
-import { BuiltInRecordingList } from './types';
+import { BuiltInRecordingList, RecordingPageState } from './types';
 
 import './show-details.scss';
-import { RecordingPageState } from '.';
 
 interface ProjectFile extends RecordingProject {
   name: string;
