+++ src/renderer/components/configure-recording-set-page/dvcfg.tsx
@@ -4,7 +4,7 @@ import { CSSTransition } from 'react-transition-group';
 
 import { Consumer } from '../../types';
 
-import { RecordingPageState } from '.';
+import { RecordingPageState } from './types';
 
 export const Dvcfg: FC<{
   setRecordingSetState: Consumer<RecordingPageState>;
